fix: update getInscription to store in new format
The original dataset was polluted with both inscription IDs and inscription numbers. In addition, the dat was sorted alphabetically in the key list, making it difficult to see the numbers in sequence without querying the entire dataset (currently 450k+ keys).

This makes it so each inscription is stored as inscription-NUM, which will allow filtering through the KV keylist to help reduce the required number of queries and make data manipulation easier.

Note: this will populate the V2 list and be available in development env to assist with migration.